WildlifeCompliance: REFACTOR fee calculation process into ApplicationService.
